TG-1712 Fix: Show core dump to display uncompressed core files
Root Cause :
Systemd coredump utility is used to generate, compress and store the coredump.
Additionaly it will also store information about the core dump in the extended
attributes of the core file.  This extended attributes are used by both
coredumpctl as well as the show core dump to display the core dump details.  Due
to Out Of Memory issue, systemd-coredump utility is not able to complete
compression of coredump, it errors out leaving a un compressed coredump.  This
un compressed coredump doesnt contain the extended attributes also.  Hence show
core dump is not able to extract any information from this uncompressed coredump
and hence didnt display the same.

Solution:
Apart from relying on the extended attributes, few information could be obtained
from the coredump file name itself.  (ie., daemon name, instance-id and
timestamp) could be fetch from core file name, where as the reason for crash
could not be fetched.
Hence the solution is enhanced to fallback to the information present of
filename if the extended attributes are not available.

Change-Id: Ief8f838104527a6359daaa2e24221e8934467718
Signed-off-by: sachin soni <sachin.soni@hpe.com>